update
修改序列号类型

+++ src/main/java/net/client/Sender.java
@@ -4,9 +4,9 @@ import com.google.protobuf.Message;
 import net.message.TCPMessage;
 
 public interface Sender {
-	void sendMessage(int msgId, Message msg, long sequence);
+	void sendMessage(int msgId, Message msg, int sequence);
 
 	void sendMessage(TCPMessage msg);
 
-	void sendMessage(int clientId, int msgId, int mapId, Message msg, long sequence);
+	void sendMessage(int clientId, int msgId, int mapId, Message msg, int sequence);
 }
