+++ src/main/java/net/message/TCPMaker.java
@@ -8,11 +8,11 @@ public class TCPMaker {
 	public TCPMaker() {
 	}
 
-	public TCPMessage wrap(int msgId, Message msg, long sequence) {
+	public TCPMessage wrap(int msgId, Message msg, int sequence) {
 		return TCPMessage.newInstance(msgId, msg.toByteArray(), sequence);
 	}
 
-	public TCPMessage wrap(int clientId, int msgId, int mapId, Message msg, long sequence) {
+	public TCPMessage wrap(int clientId, int msgId, int mapId, Message msg, int sequence) {
 		return TCPMessage.newInstance(msgId, clientId, msg.toByteArray(), mapId, sequence);
 	}
 }
